Extract Python smart pointer code for use by JIT code and not just test code
