+++ Tests/test_inference.cpp
@@ -35,6 +35,7 @@
 #include "testing_util.h"
 #include <Python.h>
 #include <absint.h>
+#include <util.h>
 #include <memory>
 #include <vector>
